refactor(replay): mechanically split bundle into bounded modules (wave17 step2)
